pyproject.toml
[project]
name = "llmjam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sounddevice",
    "crepe",
    "openai",
    "requests",
    "python-rtmidi",
    "python-dotenv>=1.1.0",
    "tensorflow-macos>=2.16.2",
    "tensorflow-metal>=1.2.0",
    "tiktoken>=0.9.0",
]
